+++ src/getEnvVar.js
@@ -11,9 +11,9 @@
 function getEnvVar(key, defaultValue) {
   if (typeof process !== 'undefined' && process.env && process.env[key]) {
     return process.env[key]
-  } else if (typeof window !== 'undefined' && window[key] != null) {
+  } else if (typeof window !== 'undefined' && window[key]) {
     return window[key]
-  } else if (typeof global !== 'undefined' && global[key] != null) {
+  } else if (typeof global !== 'undefined' && global[key]) {
     return global[key]
   } else {
     return defaultValue !== undefined ? defaultValue : null